factor out strutural tests
